feat: add navigation using coordinators continued
