+++ iMovies/Feature/Login/LoginView.swift
@@ -1,0 +1,25 @@
+//
+//  LoginView.swift
+//  iMovies
+//
+//  Created by cefalo on 31/1/24.
+//
+
+import SwiftUI
+
+struct LoginView: View {
+    
+    var body: some View {
+        ZStack {
+            Color(.systemBackground)
+            
+            
+        }
+    }
+}
+
+struct LoginView_Previews: PreviewProvider {
+    static var previews: some View {
+        LoginView()
+    }
+}
