--- iMovies/iMoviesApp.swift
@@ -1,18 +1,0 @@
-////
-////  iMoviesApp.swift
-////  iMovies
-////
-////  Created by cefalo on 24/12/23.
-////
-//
-//import SwiftUI
-//
-//@main
-//struct iMoviesApp: App {
-//
-//    var body: some Scene {
-//        WindowGroup {
-//            DashboardCoordinator()
-//        }
-//    }
-//}
